+++ scripts/ci/opam-release.sh
@@ -1,6 +1,16 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 set -e
+
+# Checks if running in dry-mode
+for arg in "$@"; do
+  case $arg in
+  "--dry-run")
+    dry_run="--dry-run"
+    echo "Running in dry-run mode. The opam packages won't be actually released."
+    ;;
+  esac
+done
 
 ci_dir="$(cd "$(dirname "$0")" && echo "$(pwd -P)/")"
 script_dir="$(dirname "$ci_dir")"
@@ -8,12 +18,12 @@ script_dir="$(dirname "$ci_dir")"
 opam_repository_fork="git@github.com:tezos/opam-repository"
 opam_dir="opam-repository"
 
-log () {
-    echo '\e[1m'"$1"'\e[0m'
+log() {
+  printf '\e[1m%s\e[0m' "$1"
 }
 
-# shellcheck source=./scripts/ci/release.sh
-. "$ci_dir/release.sh"
+# shellcheck source=./scripts/ci/octez-release.sh
+. "$ci_dir/octez-release.sh"
 
 # set up ssh credentials to access github
 mkdir -p "$HOME/.ssh"
@@ -27,24 +37,30 @@ log "Done setting up credentials."
 # call opam-release.sh with the correct arguments
 echo "$script_dir/opam-release.sh" \
   "$opam_release_tag" \
-  "https://gitlab.com/tezos/tezos/-/archive/$CI_COMMIT_TAG/$gitlab_package_name.tar.gz" \
-  "$opam_dir"
+  "https://gitlab.com/tezos/tezos/-/archive/$CI_COMMIT_TAG/$gitlab_octez_source_package_name.tar.gz" \
+  "$opam_dir" \
+  "$dry_run"
 
 "$script_dir/opam-release.sh" \
   "$opam_release_tag" \
-  "https://gitlab.com/tezos/tezos/-/archive/$CI_COMMIT_TAG/$gitlab_package_name.tar.gz" \
-  "$opam_dir"
+  "https://gitlab.com/tezos/tezos/-/archive/$CI_COMMIT_TAG/$gitlab_octez_source_package_name.tar.gz" \
+  "$opam_dir" \
+  "$dry_run"
 
 # Matches the corresponding variable in /scripts/opam-release.sh.
 branch_name="octez-$(echo "$opam_release_tag" | tr '~' -)"
 
-log "While we're here, update master on the fork..."
-cd "$opam_dir"
-git remote add github "$opam_repository_fork"
-git push github master:master
+if [ -z "$dry_run" ]; then
 
-log "Pushing $branch_name to $opam_repository_fork..."
-git push --force-with-lease github "${branch_name}:${branch_name}"
+  log "While we're here, update master on the fork..."
+  cd "$opam_dir"
+  git remote add github "$opam_repository_fork"
+  git push github master:master
 
-log "Create the pull request at:"
-log "https://github.com/ocaml/opam-repository/compare/master...tezos:opam-repository:${branch_name}"
+  log "Pushing $branch_name to $opam_repository_fork..."
+  git push --force-with-lease github "${branch_name}:${branch_name}"
+
+  log "Create the pull request at:"
+  log "https://github.com/ocaml/opam-repository/compare/master...tezos:opam-repository:${branch_name}"
+
+fi
